Restructure project and refactor
